Create shortcut for importing xlsx files
This may prove useful as we attempt to make this more user-friendly for people in the future. Ultimately, I would like to have a "drag-n-drop" sort of setup. User can save their xlsx file in a particular folder, that document is converted to a csv, then the code I have written will clean and pull the pertinent information based on my parameters.